backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
akshare>=1.18.0
tushare>=1.4.0
lightgbm>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
apscheduler>=3.10.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
scikit-learn>=1.3.0
